requirements.txt
scikit-learn>=0.22
scipy>=1.0
numba>=0.46
pot>=0.6
umap-learn>=0.3.10
joblib